requirements.txt
discord
tzdata
async-lru
tortoise-orm[asyncpg]